requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
supabase==2.3.1
python-dateutil==2.8.2
requests==2.31.0
psycopg2-binary==2.9.9
